requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
playwright==1.40.0
playwright-stealth==1.0.5
PyMuPDF==1.23.8
langchain==0.1.0
langchain-community==0.3.24
ratelimit==2.2.1
transformers==4.36.2
torch==2.1.2
lxml==4.9.3
flask==3.0.2
flask-cors==4.0.0
huggingface-hub==0.19.4 